# Add fast path for macro with single token